Chore: Update pre-commit issue link requirement
The Release Engineering Jira project is no longer publicly accessible.
As we currently require issue links for commits to lftools, this
switches to using the GitHub issue tracker for links.

Links may also use the optional 'Fixes' message identifier to cause GH
to close an issue if it is still open.

Issue: Fixes lfit/releng-lftools#10
Change-Id: I13b7592248f0d77a576d394a5a4415209777ae87
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>