CI: Add environment static credentials for next tag/push
The lftools package isn't found on test.pypi.org, which is resulting in the  fallback mechanism being used by our pypi-publish-action, which requires either a 1Password service account, and a vault credential, or a static password provided in the development/production environments. This change will pass the missing secrets through to the action.

Change-Id: I1a42ac19fb2be00740023072531944931f51d138
Signed-off-by: ModeSevenIndustrialSolutions <mwatkins@linuxfoundation.org>